requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
requests==2.31.0
moviepy==1.0.3
pandas==2.1.4
